Модуль 7. Задание 1
