+++ src/sort-type.js
@@ -1,0 +1,12 @@
+/**
+ * @fileoverview Список доступных видов сортировки
+ */
+
+'use strict';
+
+/** @enum {number} */
+module.exports = {
+  'NEW': 'new',
+  'DISCUSSED': 'discussed',
+  'POPULAR': 'popular'
+};
